Move ExampleModule to examples.module package
